feat: Update the upgrade script

+++ geek/.geek/updates.sh
@@ -8,16 +8,28 @@ case "${unameOut}" in
     MINGW*)     machine=MinGw;;
     *)          machine="UNKNOWN:${unameOut}"
 esac
+distoOut="$(lsb_release -si)"
+case "${distoOut}" in
+    Ubuntu*)    distro=Ubuntu;;
+    Pop*)       distro=PopOS;;
+    Debian*)    distro=Debian;;
+    *)          distro=""
+esac
 
 # Base commands
-alias deb-install="sudo dpkg -i $1 && sudo apt install -f"
-
+if [[ "$distro" =~ ^(Ubuntu|PopOS|Debian)$ ]]; then
+    alias deb-install="sudo dpkg -i $1 && sudo apt install -f"
+fi
 
 update() {
     if [[ "$machine" == "Linux" ]]; then
-        sudo apt update -y 
-        sudo apt upgrade -y
-    elif [[ "$machine" == "Darwin" ]]; then
+        if [[ "$distro" =~ ^(Ubuntu|PopOS|Debian)$ ]]; then
+            sudo apt update -y 
+        else 
+            echo "Unknown Distro / Package Manager"
+        fi
+
+    elif [[ "$machine" == "Mac" ]]; then
         brew update
     else
         echo "Unknown OS: $machine"
@@ -26,16 +38,38 @@ update() {
 
 upgrade() {
     if [[ "$machine" == "Linux" ]]; then
-        sudo apt upgrade -y 
-        sudo apt dist-upgrade -y
-        sudo apt autoremove -y 
-        sudo apt clear -y
+        if [[ "$distro" =~ ^(Ubuntu|PopOS|Debian)$ ]]; then
+            echo "Updating apt..."
+            sudo apt upgrade -y 
+            sudo apt dist-upgrade -y
+            sudo apt autoremove -y 
+            sudo apt autoclear -y
+        else 
+            echo "Unknown Distro / Package Manager"
+        fi
+        
+        # Flatpak
+        if [[ -x "$(flatpak)" ]]; then
+            flatpak update
+        fi
     elif [[ "$machine" == "Darwin" ]]; then
+        echo "Updating brew..."
         brew update 
         brew upgrade 
+        # Generate a lock file
+        brew bundle dump --force
         brew cleanup
     else
         echo "Unknown OS: $machine"
+    fi
+
+    if [[ -d "$HOME/.cargo/" ]]; then
+        echo "Updating rust..."
+        rustup update
+    fi
+    if [[ -d "$HOME/.nvm/" ]]; then
+        echo "Updating node..."
+        nvm install node
     fi
 }
 
@@ -44,9 +78,11 @@ install() {
     update 
 
     if [[ "$machine" == "Linux" ]]; then
-        sudo apt install -y $*
+        if [[ "$distro" =~ ^(Ubuntu|PopOS|Debian)$ ]]; then
+            sudo apt install -y "$*"
+        fi
     elif [[ "$machine" == "Darwin" ]]; then
-        brew install $@
+        brew install "$*"
     else
         echo "Unknown OS: $machine"
     fi
@@ -54,9 +90,11 @@ install() {
 
 search() {
     if [[ "$machine" == "Linux" ]]; then
-        sudo apt-cache search $@
+        if [[ "$distro" =~ ^(Ubuntu|PopOS|Debian)$ ]]; then
+            sudo apt-cache search "$*"
+        fi
     elif [[ "$machine" == "Darwin" ]]; then
-        brew search $@
+        brew search "$*"
     else
         echo "Unknown OS: $machine"
     fi
